requirements.txt
fastapi
uvicorn
python-dotenv
line-bot-sdk
supabase
